fix: changed to int instead of float

+++ main.py
@@ -27,7 +27,7 @@ def callback(indata, frames, time, status):
     q.put(indata[:])
 
 try:
-    with sd.RawInputStream(samplerate=16000, blocksize=8000, dtype="float32",
+    with sd.RawInputStream(samplerate=16000, blocksize=8000, dtype="int16",
                        channels=1, callback=callback):
         print("Listening... Speak now.")
         while True:
